+++ Packages/com.unity.inputsystem/InputSystem/Editor/AssetEditor/InputActionPropertiesView.cs
@@ -47,10 +47,6 @@ namespace UnityEngine.InputSystem.Editor
 
             if (EditorGUI.EndChangeCheck())
             {
-                // When changing the action type, clear out the expected control type.
-                if ((InputActionType)m_SelectedActionType != (InputActionType)m_ActionTypeProperty.intValue)
-                    m_SelectedControlType = 0;
-
                 if ((InputActionType)m_SelectedActionType == InputActionType.Button)
                     m_ExpectedControlTypeProperty.stringValue = "Button";
                 else if (m_SelectedControlType == 0)
@@ -60,6 +56,7 @@ namespace UnityEngine.InputSystem.Editor
 
                 m_ActionTypeProperty.intValue = (int)(InputActionType)m_SelectedActionType;
                 m_ActionTypeProperty.serializedObject.ApplyModifiedProperties();
+                UpdateProcessors(m_ExpectedControlTypeProperty.stringValue);
 
                 onChange(k_PropertiesChanged);
             }